requirements.txt
Flask==2.3.2  
Flask-SQLAlchemy==3.0.3  
Flask-Migrate==4.0.4  
Flask-Marshmallow==0.15.0  
marshmallow-sqlalchemy==0.29.0  
Flask-Cors==4.0.0  
cryptography==41.0.3  
python-dotenv==1.0.0  
gunicorn==20.1.0